% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{add_total_pool}
\alias{add_total_pool}
\title{add_total_pool}
\usage{
add_total_pool(model = NULL, decomp_list = NULL, verbose = FALSE)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
a \code{list} of 3 \code{data.frame}'s representing the variable and category decomposition, and the fitted values.
}
\description{
Add an aggregated decomposition.
}
\details{
When running a pooled model, it might be desirable to view the output of the \code{decomping()} function as an aggregate of all pools.
}
