% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{add_total_pool_to_data}
\alias{add_total_pool_to_data}
\title{add_total_pool_to_data}
\usage{
add_total_pool_to_data(data, pool_var, id_var)
}
\arguments{
\item{data}{A \code{data.frame} containing the pool and id variables provided.}

\item{pool_var}{A string representing the variable name of the pool variable.}

\item{id_var}{A string representing the variable name of the id variable.}
}
\value{
a \code{data.frame} with additional observations.
}
\description{
Add an aggregated set of observations to a \code{data.frame}
}
\details{
Add an aggregated set of observations to a \code{data.frame} based on a "pool" variable provided.
}
