% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{available_countries}
\alias{available_countries}
\title{available_countries}
\usage{
available_countries()
}
\value{
\code{data.frame} of country codes and names
}
\description{
Available countries for the economy variables
}
\details{
The \code{get_economy} function accepts country codes and/or names.
These are passed to that function as country_code or in the pool_var.
\code{available_countries} shows which country codes and/or names are accepted.
}
