% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{build_formula}
\alias{build_formula}
\title{build_formula}
\usage{
build_formula(dv, ivs, model_table = NULL)
}
\arguments{
\item{dv}{string of dependent variable name}

\item{ivs}{character vector of independent variable names}

\item{model_table}{\code{tibble}/ \code{data.frame} as created in the \code{build_model_table} function}
}
\value{
a \code{formula} object
}
\description{
Build a formula (e.g. y ~ x1 + x2)
}
\details{
Build a formula (e.g. y ~ x1 + x2) based on a dependent variable name and a model table or independent variables' names' vector
}
