% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{build_model_table}
\alias{build_model_table}
\title{build_model_table}
\usage{
build_model_table(ivs, trans_df = NULL, ts = TRUE)
}
\arguments{
\item{ivs}{character vector of variables}

\item{trans_df}{\code{data.frame} defining the non-linear transformations to apply}

\item{ts}{boolean to specify if time-series or not}
}
\value{
\code{tibble} of model table
}
\description{
Build an empty model table
}
\details{
Build an empty table as a template to capture model predictors, and transformation parameters
}
\examples{
build_model_table(c('x1','x2'))
build_model_table(colnames(mtcars))
}
