% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{check_currency}
\alias{check_currency}
\title{check_currency}
\usage{
check_currency(data, currencies = "$,£")
}
\arguments{
\item{currencies}{a comma-separated string of symbols}

\item{text}{Code to run as string}
}
\value{
parsed data.frame
}
\description{
check and convert to numeric
}
\details{
Given a comma-separated string of symbols (e.g.  "$,£"), columns are converted to numeric.
Only columns containing the symbols are checked.
Use \code{linea::check_numeric} for a more automated solution.
}
