% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_file.R
\name{check_model_file}
\alias{check_model_file}
\title{check_model_file}
\usage{
check_model_file(model_file, verbose = FALSE, return_list = TRUE)
}
\arguments{
\item{model_file}{File path to the model file as string}

\item{verbose}{A boolean to specify whether to print warnings}

\item{return_list}{Boolean to specify whether to print warnings}
}
\value{
Messages and the checked file
}
\description{
Check the excel model file
}
\details{
Check the model file contains all the needed sheets.
}
