% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{check_model_table}
\alias{check_model_table}
\title{check_model_table}
\usage{
check_model_table(model_table, verbose = FALSE, trans_df = NULL)
}
\arguments{
\item{model_table}{\code{tibble}/ \code{data.frame} as created in the \code{build_model_table} function}

\item{verbose}{A boolean to specify whether to print warnings}

\item{trans_df}{\code{data.frame} defining the non-linear transformations to apply}
}
\value{
\code{tibble} of model table
}
\description{
Check \code{model_table}
}
\details{
Check the columns of a \code{model_table} and generate any missing ones where possible.
}
