% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{check_numeric}
\alias{check_numeric}
\title{check_numeric}
\usage{
check_numeric(data, guess = TRUE)
}
\arguments{
\item{data}{data.frame to check}

\item{guess}{a boolean to specify weather to guess data-type}
}
\value{
parsed data.frame
}
\description{
check_numeric using \code{parsr}
}
\details{
An implementation of \code{readr::parse_number} which will look for numbers to parse, or alternatively guess using \code{readr::parse_guess}.
This process is applied to too all columns of the input data.frame \code{data}.
}
