% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{check_trans_df}
\alias{check_trans_df}
\title{check_trans_df}
\usage{
check_trans_df(trans_df)
}
\arguments{
\item{trans_df}{\code{data.frame} defining the non-linear transformations to apply}
}
\value{
\code{data.frmae} of trans_df
}
\description{
Check trans_df based on default_trans_df
}
\details{
Check that the trans_df \code{data.frame} contains all the necessary columns.
If not, create them and return the amended trans_df.
}
\examples{
default_trans_df() \%>\% check_trans_df()
}
