% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{check_ts}
\alias{check_ts}
\title{check_ts}
\usage{
check_ts(data, date_col, allow_non_num = T, verbose = F)
}
\arguments{
\item{data}{The dataframe containing the column specified}

\item{date_col}{The date column name as a string}

\item{allow_non_num}{A boolean to specify whether to include only date and numeric columns}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{data.frame}
}
\description{
Check time series dataframe
}
\details{
Check if dataframe contains specified date column and that its data-type
}
