% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{color_palette}
\alias{color_palette}
\title{color_palette}
\usage{
color_palette()
}
\value{
character vector of colors in hexadecimal notation
}
\description{
A pre-loaded color palette
}
\details{
A pre-loaded color palette that can be used in charting functions
}
