% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_next_step.R
\name{combo_number}
\alias{combo_number}
\title{combo_number}
\usage{
combo_number(model = NULL, trans_df = NULL, verbose = FALSE)
}
\arguments{
\item{model}{Model object}

\item{trans_df}{\code{data.frame} containing the transformations, variables and parameter values}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
numeric count of what_combo models
}
\description{
The number of models from all combinations of transformations and variables
}
\details{
Count the number of models from all combinations of transformations and variables
}
\examples{

# using a model object
data = linea::sales_ts
dv = 'sales'
ivs = c("relative_price","vod_spend", "dec", "mar")


trans_df = data.frame(
  name = c('diminish', 'decay', 'hill', 'exp'),
  ts = c(FALSE,TRUE,FALSE,FALSE),
  func = c(
    'linea::diminish(x,a)',
    'linea::decay(x,a)',
    "linea::hill_function(x,a,b,c)",
    '(x^a)'
  ),
  order = 1:4
) \%>\%
  dplyr::mutate(display_spend = dplyr::if_else(condition = name == 'hill',
                                               '(1,50),(1),(1,100)',
                                               '')) \%>\%
  dplyr::mutate(display_spend = dplyr::if_else(condition = name == 'decay',
                                              '.1,.7 ',
                                              display_spend))

model = run_model(data = data,dv = dv,ivs = ivs, trans_df = trans_df)

combos = combo_number(model = model,trans_df = trans_df)

#using the trans_df
combo_number(trans_df = trans_df)
}
