% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{countries}
\format{
A tibble representing countries as per the \code{wbstats::wb_countries()}
}
\source{
World Bank \code{wbstats::wb_countries()} (https://data.worldbank.org/country)
}
\usage{
countries
}
\description{
dataset of world bank countries
}
\keyword{datasets}
