% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{decay}
\alias{decay}
\title{decay}
\usage{
decay(v, decay)
}
\arguments{
\item{v}{numeric vector}

\item{decay}{The rate of decay as a numeric decimal}
}
\value{
The transformed vector \code{v}
}
\description{
Time series decay
}
\details{
Applies the specified decay on the input vector, v
}
\examples{
decay(c(1,0,0,0,1,0,0,0,2), 0.5)
decay(c(1,0,0,0,1,0,0,0,2), 0.1)
}
