% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{default_trans_df}
\alias{default_trans_df}
\title{default_trans_df}
\usage{
default_trans_df()
}
\value{
\code{data.frame} containing the transformations
}
\description{
The default transformation \code{data.frame}
}
\details{
The default transformation \code{data.frame} containing the functions \code{decay}, \code{diminish}, \code{lag}, and \code{ma}
}
