% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{diminish}
\alias{diminish}
\title{Diminish}
\usage{
diminish(v, m, abs = TRUE)
}
\arguments{
\item{v}{Numeric vector}

\item{m}{Scale of diminishing as a numeric integer or decimal}

\item{abs}{Boolean to determine if diminishing scale \code{m} is a percentage or absolute value}
}
\value{
The transformed vector \code{v}
}
\description{
Negative exponential (Diminish returns)
}
\details{
Applies the negative exponential (1 - exp(-x/m)) on the input vector
}
\examples{
diminish(c(1,0,0,0,1,0,0,0,2), 1)
diminish(c(1,0,0,0,1,0,0,0,2), 1, FALSE)
}
