% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_file.R
\name{export_model}
\alias{export_model}
\title{export_model}
\usage{
export_model(model, path = NULL, overwrite = FALSE)
}
\arguments{
\item{model}{Model object}

\item{path}{File path to the model file as string}

\item{overwrite}{Boolean to specify whether to overwrite the file in the specified path}
}
\description{
Export model to excel file
}
\details{
Export a model to an excel file
}
