% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{filter_decomp_pool}
\alias{filter_decomp_pool}
\title{filter_decomp_pool}
\usage{
filter_decomp_pool(decomp, pool, verbose = TRUE)
}
\arguments{
\item{decomp}{A \code{list} of \code{data.frame} from a model object or generated using \code{decomping}.}

\item{pool}{A string representing the variable name of the pool variable.}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
a \code{list} of 3 \code{data.frame}'s representing the variable and category decomposition, and the fitted values.
}
\description{
Filter a model's decomposition based on a given pool.
}
\details{
Filter all \code{data.frame}'s within a model's decomposition based on a given pool.
}
