% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{fit_chart}
\alias{fit_chart}
\title{fit_chart}
\usage{
fit_chart(
  model = NULL,
  decomp_list = NULL,
  pool = NULL,
  verbose = F,
  colors = NULL
)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{verbose}{A boolean to specify whether to print warnings}

\item{colors}{character vector of colors in hexadecimal notation}
}
\description{
Dependent Variable, Predictions and Residuals Line Chart
}
\details{
Plot the dependent variable, predictions and Residuals as a line chart over the id variable which can be supplied to the \code{decomping} function.
}
