% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{get_oecd_data}
\alias{get_oecd_data}
\title{get_oecd_data}
\usage{
get_oecd_data(
  data,
  country_code = NULL,
  date_col_name = NULL,
  date_type = NULL,
  append = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} containing data for analysis}

\item{country_code}{A string indicating which county's economy data to retrieve.}

\item{date_col_name}{The date column name as a string}

\item{append}{A boolean to specify whether to return only the economy variables or the original input data plus the economy variables}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{data.frame} with added variables
}
\description{
Generate economy variables using OECD data
}
\details{
Generate economy variables from a \code{data.frame} containing a date-type variable.
The function will collect data from the OECD, based on the country_code input.
The indicators that will be gathered are quarterly GDP (index) and monthly unemployment.
Appropriate country codes can be viewed using the tibble \code{linea::countries$iso2c}.
}
\examples{
linea::sales_ts \%>\%
   get_oecd_data(
       date_col_name = 'week',
       country_code = 'US',
       append=FALSE,
       date_type="weekly starting")
}
