% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{get_offset}
\alias{get_offset}
\title{get_offset}
\usage{
get_offset(data, model_table)
}
\arguments{
\item{data}{\code{data.frame} containing variables included in the model_table}

\item{model_table}{\code{tibble}/ \code{data.frame} as created in the \code{build_model_table} function}
}
\value{
a \code{matrix} object
}
\description{
Build a formula's offset
}
\details{
Build a formula's offset (e.g. lm(y ~ x1, offset = x2*-4)) dynamically
based no the inputs of the model_table.
}
