% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_next_step.R
\name{get_vector_from_str}
\alias{get_vector_from_str}
\title{get_vector_from_str}
\usage{
get_vector_from_str(string, sep = ",", zero = TRUE)
}
\arguments{
\item{string}{a string containing separated by a separator}

\item{sep}{a string representing the separator}

\item{zero}{a boolean defining whether the output vector must contain a zero}
}
\value{
numeric vector
}
\description{
get a numeric vector from string (e.g. '1,2,3')
}
\details{
Get a numeric vector from string containing numbers separated by a separator (e.g. '1,2,3').
}
\examples{
get_vector_from_str('1,2,3')
get_vector_from_str('0.1')
get_vector_from_str('1;2;3',sep=';')
}
