% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_file.R
\name{import_model}
\alias{import_model}
\title{import_model}
\usage{
import_model(path, verbose = F)
}
\arguments{
\item{path}{File path to the model file as string}

\item{verbose}{Boolean to specify whether to return the checked file}
}
\value{
model object
}
\description{
Import and run the excel model file
}
\details{
Import and run the excel model file using \code{check_model_file} and \code{run_model} functions
}
