% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{is_daily}
\alias{is_daily}
\title{is_daily}
\usage{
is_daily(dates)
}
\arguments{
\item{dates}{a date-type or numeric vector}
}
\value{
boolean to specify whether the time series has a daily frequency
}
\description{
Check if a time-series is daily
}
\details{
Check if a time-series is daily return boolean
}
