% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{is_uniform_ts}
\alias{is_uniform_ts}
\title{is_uniform_ts}
\usage{
is_uniform_ts(dates)
}
\arguments{
\item{dates}{a date-type or numeric vector}
}
\value{
boolean
}
\description{
Check if a time-series is uniform
}
\details{
Check if a time-series is uniform, where the step (e.g. days(1),weeks(7)) is consistent
}
