% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{is_weekly}
\alias{is_weekly}
\title{is_weekly}
\usage{
is_weekly(dates)
}
\arguments{
\item{dates}{a date-type or numeric vector}
}
\value{
boolean
}
\description{
Check if a time-series is weekly
}
\details{
Check if a time-series is weekly return boolean
}
