% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{lag}
\alias{lag}
\title{Lag}
\usage{
lag(v, l, strategy = "extremes")
}
\arguments{
\item{v}{Numeric vector}

\item{l}{Lag as an integer}

\item{strategy}{string to determine the NAs generated by the lag should be filled with zeros or with the extremities' values}
}
\value{
The lagged vector \code{v}
}
\description{
Lag by \code{l}
}
\details{
Applies a lag of \code{l} on the input vector
}
\examples{
lag(c(1,0,0,0,1,0,0,0,2), 1)
lag(c(1,0,0,0,1,0,0,0,2), -2)
lag(c(1,0,0,0,1,0,0,0,2), -2, strategy = 'zero')
}
