% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{ma}
\alias{ma}
\title{MA}
\usage{
ma(v, width, align = "center", zero = TRUE)
}
\arguments{
\item{v}{Numeric vector}

\item{width}{Width of moving average window as an integer}

\item{align}{Either string "center", "left", or "right"}

\item{zero}{Boolean to determine the NAs generated by the moving average should be filled with zeros or with the vector's mean.}
}
\value{
The modified vector \code{v}
}
\description{
Moving Average
}
\details{
Applies a moving average on the input vector. The type of moving average is defined by the argument \code{align}.
}
\examples{
ma(c(1,0,0,0,1,0,0,0,2), 3)
ma(c(1,0,0,0,1,0,0,0,2), 3, align = "right")
ma(c(1,0,0,0,1,0,0,0,2), 3, zero = FALSE)
}
