% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{parsr}
\alias{parsr}
\title{parsr}
\usage{
parsr(v, guess = TRUE)
}
\arguments{
\item{v}{value(s) to parse}

\item{guess}{a boolean to specify weather to guess data-type}
}
\value{
parsed value(s)
}
\description{
an implementation of \code{readr::parse_number}
}
\details{
an implementation of \code{readr::parse_number} which will look for numbers to parse, or alternatively guess using \code{readr::parse_guess}
}
