% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pooled_gt_data}
\alias{pooled_gt_data}
\title{pooled_gt_data}
\format{
A tibble representing google trends variables over time.
}
\source{
Google Trends (https://trends.google.com/)
}
\usage{
pooled_gt_data
}
\description{
example pooled dataset
}
\keyword{datasets}
