% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{read_xcsv}
\alias{read_xcsv}
\title{read_xcsv}
\usage{
read_xcsv(file, sheet = NULL, verbose = F)
}
\arguments{
\item{file}{The file path that points to either a csv or excel file ending in csv, xls, xlsx, or xlsm}

\item{sheet}{For excel files, the sheet name as a string or number as an integer}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{data.frame}
}
\description{
Reads flat files: either csv or excel
}
\details{
Reads csv or excel files with the suffixes csv, xls, xlsx, xlsm
}
\examples{
read_xcsv("https://raw.githubusercontent.com/paladinic/data/main/ecomm_data.csv")  
}
