% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_next_step.R
\name{run_next_model}
\alias{run_next_model}
\title{run_next_model}
\usage{
run_next_model(model, test, results_row = 1)
}
\arguments{
\item{model}{Model object}

\item{test}{output of \code{linea::what_next()} function}

\item{results_row}{numeric value of the model (i.e. row from what_combo()$results) to ru}
}
\value{
model object
}
\description{
generate the model object from the output of \code{linea::what_next()}
}
\details{
Generate the mode object from the output of \code{linea::what_next()}
Using the specs from the output of \code{linea::what_next()} a new model is run.
}
