% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{run_text}
\alias{run_text}
\title{run_text}
\usage{
run_text(text, env)
}
\arguments{
\item{text}{Code to run as string}

\item{env}{environment object specifying the environment}
}
\value{
text expression output
}
\description{
run text as R code
}
\details{
Run a text string as R code
}
