% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{TRY}
\alias{TRY}
\title{TRY}
\usage{
TRY(x, verbose = FALSE)
}
\arguments{
\item{x}{The expression to try}

\item{verbose}{boolean to specify whether to print the error}
}
\value{
expression output or \code{NULL}
}
\description{
TRY or NULL
}
\details{
A tryCatch implementation that returns NULL when an error is thrown
}
