% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{vapply_transformation}
\alias{vapply_transformation}
\title{vapply_transformation}
\usage{
vapply_transformation(v, trans_df = NULL, verbose = FALSE)
}
\arguments{
\item{v}{Numeric vector to be transformed}

\item{trans_df}{\code{data.frame} defining the non-linear transformations to apply}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
Transformed numeric vector
}
\description{
Transform vector based on transformation parameters
}
\details{
Transform vector based on the transformation parameters of the trans_df
}
