% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_data}
\alias{wb_data}
\title{wb_data}
\format{
A tibble representing macroeconomic indicators
}
\source{
World Bank \code{wbstats::wb_data()} (https://data.worldbank.org/country)
}
\usage{
wb_data
}
\description{
dataset from the World Bank representing the macroeconomic indicators:
GDP per capita, Unemployment, and Population.
}
\keyword{datasets}
