% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{wtf_chart}
\alias{wtf_chart}
\title{wtf_chart}
\usage{
wtf_chart(
  model,
  start = NULL,
  text_type = "value",
  text_rounding = 1,
  title = "Waterfall Chart",
  colors = "viridis",
  plot_bgcolor = "rgba(0, 0, 0, 0)",
  paper_bgcolor = "rgba(0, 0, 0, 0)",
  totals_color = "#1c0022",
  font_color = "#1c0022",
  zero_line_color = "#1c0022",
  grid_line_color = "#1c0022",
  verbose = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{start}{a date object specifying the start of the year-on-year comparison period}

\item{text_type}{a character string specifying how text should be displayed. Either "K", "M", or "value".}

\item{text_rounding}{a numeric value specifying how many decimals should be included in the text.}

\item{title}{a character string specifying the title of the chart}

\item{plot_bgcolor}{string representing chart color of plot_bgcolor}

\item{paper_bgcolor}{string representing chart color of paper_bgcolor}

\item{font_color}{string representing chart color of font}

\item{zero_line_color}{string representing chart color zerolinecolor}

\item{grid_line_color}{string representing chart color of gridcolor}

\item{verbose}{A boolean to specify whether to print warnings}

\item{pool}{string specifying a group within the pool column to be filtered}
}
\value{
a \code{plotly} line chart of the model's response curves
}
\description{
Waterfall chart of categories from a model object
}
\details{
Waterfall chart of categories from a model object, representing the changes in categories year-on-year
}
