% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{acf_chart}
\alias{acf_chart}
\title{acf_chart}
\usage{
acf_chart(model = NULL, decomp_list, pool = NULL, color = "black", verbose = F)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{color}{string specifying bar color}

\item{verbose}{A boolean to specify whether to print warnings}
}
\description{
Bar chart of autocorrelation function
}
\details{
A bar chart meant to assess the correlation of the residuals with lagged versions of themselves.
}
