% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{apply_normalisation}
\alias{apply_normalisation}
\title{apply_normalisation}
\usage{
apply_normalisation(
  raw_data = NULL,
  meta_data = NULL,
  model_table = NULL,
  dv = NULL,
  verbose = F
)
}
\arguments{
\item{raw_data}{\code{data.frame} containing data for analysis}

\item{meta_data}{\code{data.frame} mapping variable names to their roles (i.e. POOL)}

\item{model_table}{\code{data.frame} as created in the \code{build_model_table} function}

\item{dv}{string specifying the dependent variable name}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{list} containing a \code{tibble} of normalised data and a \code{tibble} of pool means
}
\description{
Normalise data based on pool mean
}
\details{
Normalise data by dividing all values in each pool by that pool's mean
}
