% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{apply_transformation}
\alias{apply_transformation}
\title{#' apply_transformation}
\usage{
apply_transformation(
  data = NULL,
  model_table = NULL,
  trans_df = NULL,
  meta_data = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} containing data for analysis}

\item{model_table}{\code{data.frame} as created in the \code{build_model_table} function}

\item{trans_df}{\code{data.frame} defining the non-linear transformations to apply}

\item{meta_data}{\code{data.frame} mapping variable names to their roles (i.e. POOL)}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{tibble} of raw_data with added transformed variables
}
\description{
Transform data based on model table
}
\details{
Transform data based on the model table by applying the transformation functions (e.g. \code{decay}, \code{diminish}, \code{lag}, and \code{ma}) with the specified parameters to the respective variable
}
