% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{default_trans_df}
\alias{default_trans_df}
\title{default_trans_df}
\usage{
default_trans_df(ts = TRUE)
}
\arguments{
\item{ts}{boolean to specify if time-series or not}
}
\value{
\code{data.frmae} of trans_df
}
\description{
The default trans_df
}
\details{
Generate the default trans_df \code{data.frame} with functions from linea:
\itemize{
\item decay
\item diminish
\item hill_function
\item ma
\item lag
}
}
\examples{
default_trans_df()
default_trans_df(ts = TRUE)
}
