% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{fit_chart}
\alias{fit_chart}
\title{fit_chart}
\usage{
fit_chart(
  model = NULL,
  decomp_list = NULL,
  pool = NULL,
  verbose = FALSE,
  colors = NULL,
  predicted_line_color = "#00cf74",
  actual_line_color = "#1c0022",
  resid_line_color = "#c90f3a",
  plot_bgcolor = "rgba(0, 0, 0, 0)",
  paper_bgcolor = "rgba(0, 0, 0, 0)",
  font_color = "#1c0022",
  zero_line_color = "#1c0022",
  grid_line_color = "#1c0022"
)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{verbose}{A boolean to specify whether to print warnings}

\item{colors}{character vector of colors in hexadecimal notation}

\item{predicted_line_color}{string representing chart color of the predicted line}

\item{actual_line_color}{string representing chart color of the actual line}

\item{resid_line_color}{string representing chart color of residual line}

\item{plot_bgcolor}{string representing chart color of plot_bgcolor}

\item{paper_bgcolor}{string representing chart color of paper_bgcolor}

\item{font_color}{string representing chart color of font}

\item{zero_line_color}{string representing chart color zerolinecolor}

\item{grid_line_color}{string representing chart color of gridcolor}
}
\value{
a \code{plotly} line chart of the model's prediction and actual
}
\description{
Dependent Variable, Predictions and Residuals Line Chart
}
\details{
Plot the dependent variable, predictions and Residuals as a line chart over the id variable which can be supplied to the \code{decomping} function.
}
\examples{
run_model(data = mtcars,dv = 'mpg',ivs = 'cyl') \%>\% fit_chart()
}
