% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{get_economy}
\alias{get_economy}
\title{get_economy}
\usage{
get_economy(
  data,
  country_code = NULL,
  date_col_name = NULL,
  append = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} containing data for analysis}

\item{country_code}{A string indicating which county's economy data to retrieve.}

\item{date_col_name}{The date column name as a string}

\item{append}{A boolean to specify whether to return only the economy variables or the original input data plus the economy variables}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
\code{data.frame} with added variables
}
\description{
Generate economy variables
}
\details{
Generate economy variables from a \code{data.frame} containing a date-type variable.
The function will collect data from the world bank based on the country_code input.
Appropriate country codes can be viewed using the tibble \code{linea::countries$iso2c}.
}
\examples{
linea::sales_ts \%>\%
   get_economy(date_col_name = 'week',country_code = 'US',append=FALSE)
}
