% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{get_variable_t}
\alias{get_variable_t}
\title{get_variable_t}
\usage{
get_variable_t(
  model_table,
  excl_intercept = TRUE,
  excl_dup = TRUE,
  excl_blanks = FALSE,
  trans_df = NULL
)
}
\arguments{
\item{model_table}{\code{tibble}/ \code{data.frame} as created in the \code{build_model_table} function}

\item{excl_intercept}{Boolean to specify whether to drop the "(Intercept)" row of the model table}

\item{excl_dup}{Boolean to specify whether to drop the duplicated rows of the model table}

\item{excl_blanks}{Boolean to specify whether to drop the blank rows of the model}

\item{trans_df}{\code{data.frame} defining the non-linear transformations}
}
\value{
\code{tibble} of model table with added variable_t column
}
\description{
Generate more specific variable names
}
\details{
Generate variable names that capture the transformations applied
}
\examples{
build_model_table(colnames(mtcars)) \%>\%
   get_variable_t()
}
