% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{heteroskedasticity_chart}
\alias{heteroskedasticity_chart}
\title{heteroscedasticity_chart}
\usage{
heteroskedasticity_chart(
  model = NULL,
  decomp_list = NULL,
  pool = NULL,
  color = "black",
  plot_bgcolor = "rgba(0, 0, 0, 0)",
  paper_bgcolor = "rgba(0, 0, 0, 0)",
  font_color = "#1c0022",
  zero_line_color = "#1c0022",
  verbose = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{color}{string specifying bar color}

\item{plot_bgcolor}{string representing chart color of plot_bgcolor}

\item{paper_bgcolor}{string representing chart color of paper_bgcolor}

\item{font_color}{string representing chart color of font}

\item{zero_line_color}{string representing chart color zerolinecolor}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
a \code{plotly} scatter chart of the model's dependent variable over residuals
}
\description{
Scatter of Residuals over dependent Variable
}
\details{
Plot a scatter chart of residuals over the dependent variable.
This is meant to assess the consistency of the residuals' variance across the dependent variable.
}
