% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{hill_function}
\alias{hill_function}
\title{hill_function}
\usage{
hill_function(v, m, b = 1, k = 1, abs = TRUE)
}
\arguments{
\item{v}{Numeric vector}

\item{m}{Numeric integer or decimal}

\item{b}{Numeric integer or decimal}

\item{k}{Numeric integer or decimal}

\item{abs}{Boolean to determine if diminishing scale \code{m} is a percentage or absolute value}
}
\value{
The transformed vector \code{v}
}
\description{
Hill Function
}
\details{
Applies the Hill Function (1 - exp(-x)) on the input vector
}
\examples{
hill_function(c(1,0,0,0,1,0,0,0,2), m=1)
hill_function(c(1,0,0,0,1,0,0,0,2), m=1, abs = FALSE)
hill_function(c(1,0,0,0,1,0,0,0,2), m=1, k = 2)
}
