% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{resid_hist_chart}
\alias{resid_hist_chart}
\title{resid_hist_chart}
\usage{
resid_hist_chart(
  model = NULL,
  decomp_list = NULL,
  pool = NULL,
  color = "black",
  plot_bgcolor = "rgba(0, 0, 0, 0)",
  paper_bgcolor = "rgba(0, 0, 0, 0)",
  font_color = "#1c0022",
  zero_line_color = "#1c0022",
  verbose = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{decomp_list}{list object generated by the \code{decomping} function.}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{color}{string specifying bar color}

\item{plot_bgcolor}{string representing chart color of plot_bgcolor}

\item{paper_bgcolor}{string representing chart color of paper_bgcolor}

\item{font_color}{string representing chart color of font}

\item{zero_line_color}{string representing chart color zerolinecolor}

\item{verbose}{A boolean to specify whether to print warnings}
}
\value{
a \code{plotly} histogram of the model's residuals
}
\description{
Histogram of Model Residuals
}
\details{
Plot a histogram to visualise the distribution of residuals.
This is meant to assess the residual distribution's normality.
}
