% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting.R
\name{response_curves}
\alias{response_curves}
\title{response_curves}
\usage{
response_curves(
  model,
  x_min = NULL,
  x_max = NULL,
  interval = NULL,
  pool = NULL,
  trans_only = FALSE,
  colors = "viridis",
  plot_bgcolor = "rgba(0, 0, 0, 0)",
  paper_bgcolor = "rgba(0, 0, 0, 0)",
  font_color = "#1c0022",
  zero_line_color = "#1c0022",
  grid_line_color = "#1c0022",
  plotly = TRUE,
  allow_ts_trans = FALSE,
  verbose = FALSE,
  table = FALSE,
  histogram = FALSE,
  points = FALSE,
  non_zero_obs = FALSE,
  add_intercept = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{x_min}{number specifying horizontal axis min}

\item{x_max}{number specifying horizontal axis max}

\item{interval}{number specifying interval between points of the curve}

\item{pool}{string specifying a group within the pool column to be filtered}

\item{trans_only}{a boolean specifying whether to display non-linear only \code{y = b*dim_rest(x)}}

\item{colors}{character vector of colors in hexadecimal notation}

\item{plot_bgcolor}{string representing chart color of plot_bgcolor}

\item{paper_bgcolor}{string representing chart color of paper_bgcolor}

\item{font_color}{string representing chart color of font}

\item{zero_line_color}{string representing chart color zerolinecolor}

\item{grid_line_color}{string representing chart color of gridcolor}

\item{plotly}{A boolean to specify whether to include use ggplot over plotly}

\item{verbose}{A boolean to specify whether to print warnings}

\item{table}{A boolean to specify whether to return a \code{data.frame} of the response curves}

\item{histogram}{A boolean to specify whether to include the histograms from the data}

\item{points}{A boolean to specify whether to include the points from the data on the curve}

\item{non_zero_obs}{A boolean to specify whether to include observations with value zero in the points and histograms}

\item{add_intercept}{A boolean to specify whether to include the intercept whne calculating the curves}
}
\value{
a \code{plotly} line chart of the model's response curves
}
\description{
Line chart of variable response curves
}
\details{
Line chart of variable response curves visualising the relationship of each independent variable with the dependent variable
}
\examples{
model = run_model(data = mtcars,
                  dv = 'mpg',
                  ivs = c('disp','wt'),
                  pool_var='cyl',
                  normalise_by_pool=TRUE)
model \%>\%
   response_curves(pool='4')
model = run_model(data = mtcars,dv = 'mpg',ivs = c('wt','cyl','disp'))

model \%>\%
   response_curves()

run_model(data = scale(mtcars) \%>\%
              data.frame(),
          dv = 'mpg',
          ivs = c('wt','cyl','disp')) \%>\%
   response_curves(points = TRUE,x_min = -3,x_max = 3)
}
