% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_next_step.R
\name{what_next}
\alias{what_next}
\title{what_next}
\usage{
what_next(
  model = NULL,
  data = NULL,
  verbose = FALSE,
  datatable_output = FALSE,
  r2_diff = TRUE
)
}
\arguments{
\item{model}{Model object}

\item{data}{\code{data.frame} containing data for analysis}

\item{verbose}{A boolean to specify whether to print warnings}

\item{datatable_output}{A boolean to specify whether the output should be a pretty DT::datatable}

\item{r2_diff}{A boolean to determine whether to add a column to compare new and original model R2}
}
\value{
\code{data.frame} mapping variables' to the respective model's statistics.
}
\description{
run model with other variables from the data
}
\details{
Run a separate model for each numeric variable in the data provided.
Then, for each model run, return that model's fit and the variables' statistics.
}
\examples{
run_model(
    data = mtcars,
    dv = 'mpg',
    ivs = c('disp', 'cyl')
  ) \%>\% 
  what_next()
}
